pro/pyproject.toml
[project]
name = "mediacrawlerpro-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==24.1.0",
    "aiohttp>=3.12.2",
    "aiomysql==0.3.2",
    "cryptography>=45.0.3",
    "cssselect==1.1.0",
    "httpx==0.28.1",
    "loguru>=0.7.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "parsel==1.6.0",
    "pillow==10.4.0",
    "pydantic==2.5.2",
    "pyhumps==3.8.0",
    "redis==4.6.0",
    "tenacity==8.2.2",
    "typer[all]>=0.16.0",
    "pre-commit>=3.5.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
